服务

+++ DataAnalysis_Server/DataAnalysis.Component.Tools/Constant/HuoBiContract.cs
@@ -17,6 +17,6 @@ namespace DataAnalysis.Component.Tools.Constant
         /// </summary>
         public static Dictionary<string, string> topicDic = new Dictionary<string, string>();
 
-        public static Dictionary<string, double> depthDic = new Dictionary<string, double>();
+        public static Dictionary<string, decimal> depthDic = new Dictionary<string, decimal>();
     }
 }
